Libft

+++ ft_strlen.c
@@ -11,10 +11,14 @@
 /* ************************************************************************** */
 
 #include "libft.h"
+#include "libft.h"
+#include "libft.h"
+#include "libft.h"
+#include "libft.h"
 
-int	ft_strlen(char const *str)
+size_t	ft_strlen(char const *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
